rag/support-agent-weaviate/pyproject.toml
[project]
name = "nebius-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "exa-py",
    "httpx",
    "keywordsai",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langgraph",
    "llama-index-core",
    "openai",
    "pillow",
    "pydantic",
    "pypdf",
    "python-dotenv",
    "requests",
    "slack-bolt",
    "slack-sdk",
    "streamlit",
    "weaviate-client",
    "watchdog",
]
